move demo addresses to test folder
